requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.1
pymongo==4.6.1
pandas==2.2.1
python-dateutil==2.8.2
tornado>=6.2.0
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
jinja2==3.1.3 